Константы перенесены в отдельный файл
