+++ src/utils/const.js
@@ -1,0 +1,16 @@
+const FILTERS_TEXT = {
+  all: 'All movies',
+  watchlist: 'Watchlist',
+  history: 'History',
+  favorites: 'Favorites',
+};
+
+const SORT_TYPE = {
+  DEFAULT: 'default',
+  BY_DATE: 'date',
+  BY_RATING: 'rating',
+};
+
+const EMOJIES = ['smile', 'sleeping', 'puke', 'angry'];
+
+export { FILTERS_TEXT, SORT_TYPE, EMOJIES };
